mcp_server/requirements.txt
fastmcp==0.3.1
fastapi
uvicorn
requests

overpy
cdsapi
pandas
xarray

pyyaml